created first iteration of the dir-watch.py
program takes a directory as the only parameter and it monitors that directory recursively. If any file/directory is deleted, added, or modified it prints out that file/directory's path.

Tested on win7, win8.1 and debian. For some actions it will print out multiples (this is because the on_any_event method runs twice, ex. if the file was created it will run once for the creation of the file, and once for the modifying of the file)